2.0.0 rewrite.
This is essentially a full ES2015 rewrite of the module, bringing
it in line with the status of the Drupal module (it was still on the
pre-2015 version of the parser) and extending it just a bit further
to a proper lexer/parser structure.

Opening tags, closing tags and text fragments are all parser tokens now.